final submission and instructions
this is my final submission of part 1 and it includes a screenshot of my repository and commits. it also includes a notepad with my repository link. the instructions to open the porgram is as follows: open visual studio and click on open folder, open my folder "ST10091914_PROG6221_Part1," then click to open the folder "PartOne," and then click on "program.cs" the program will open and run as intended.